[2/3] wifi: Disable ADB over network on disconnect
ADB over network is not disabled before the next reboot, which creates
a potential security threat when this setting is left enabled and the
device gets re-connected to public wifi networks.

To mitigate that risk, introduce a new switch to disable ADB over network
as soon as the device gets disconnected from wifi.

Change-Id: Ie6dc2e59c08618804f4cfafd9b8879b3ec7da9bc