DLKM: Add recovery for driver state change failure
With DLKM (WIFI_DRIVER_MODULE_PATH) set, the driver state change for
WIFI_DRIVER_STATE_ON is checked after the insmod is done. If this
driver state is not set to WIFI_DRIVER_STATE_ON due to firmware not
ready, there is no recovery with the current implementation. This
leaves the driver loaded but not functional and moreover any further
attempts to load the driver fails as the driver is already loaded.
Address this scenario by removing the driver on the driver state
change failure for DLKM drivers.

Change-Id: I99e425dee0b1d59ae53e1722a137db3357f5d56e
CRs-Fixed: 2214027