Set mUseSingleRadioChainScanResults to TRUE by default.
Legacy architectures and the ones with antenna sharing design operate with
single chain and thus would need the scan results with single radio
chain. Thus, set this configuration to TRUE by default. This should also
not majorly impact the architectures with dual chain. Do not want to
configure this in XML per device / architecture.

Change-Id: I93a2bdf9efbe161cc1945370eb0976f57f0617c1
CRs-Fixed: 2236186