wifi: Skip MODE_CHANGED_ACTION before WifiController state's start
This indication from Location Manager is obtained before the
Wi-Fi Controller's first state entered and this resulted in
a FATAL EXCEPTION with the reason "The start method not called,received
msg: { when=0 what=155655 target=com.android.internal.util.
StateMachine$SmHandler }" .
This is because the device is still encrypted and thus
checkAndStartWifi has failed , but Location Service sends
this intent ( probably does not check if the device is decrypted
before sending the intent ? ).
Address this issue in Wi-Fi , by checking if wificontroller is
started before processing such intents . Set wificontroller as
started when it enters in to any of the initial states

Change-Id: I4d97a2d26db0eb28fee6fabec787ce34a0946a02
CRs-Fixed: 2233921