Wifi: Disconnect on IP_REACHABILITY_LOST for a specific period
IP_REACHABILITY_LOST shall be triggered on a NUD_FAIL indication. The
current logic to check for the IP Reachability is to wait for the
response for the configured number of ARP requests to the gateway before
triggering a NUD_FAILURE. This feature when enabled by default is resulting
into some false positives (due to the ARP responses getting delayed from
the AP), leading to a disconnection / reconnection to the network.

Considering the pros / cons of this feature, this commit customizes this
behavior to be active only for the first 10 seconds after the connect/roam
to the AP. This should help to identify the roaming across multiple subnets.
The new architectures have offloaded this IPReachabilityMonitor to the host
driver which have more control of the Wi-Fi statistics before they trigger
the disconnect. This should address any issue due to the data stalls.

All said,this enhancement of "Disconnect on IP_REACHABILITY_LOST for a
specific period" is enabled by default and controlled by the property
"persist.vendor.wifi.enableIpReachabilityMonitorPeriod" .
Set this to 0 if this enhancement is not required.

Change-Id: I64cbf358e8281556b5a02df26a62396996f88aca
CRs-Fixed: 2116615