wifi: Broadcast p2p disable info on interface destroy.
This address the scenario where the P2P interface is destroyed
by the Wi-Fi State machine when SoftAP is ON (due to P2P and SoftAP
being mutually exclusive), but not indicated to the applications
on this status(wfd does not know this). This would result in wfd
information not getting set again(after the p2p restart - SoftAp off)
and thus the subsequent scan operations does not do the p2p scan's
with wfd ie's. Thus, indicate such notifications on destroy to
the applications.

CRs-Fixed: 2245215
Change-Id: I555a5520dbe86b147581b4d69bcbc8359e4cf293