Consider the STA connections on SAP+SAP while tearing down
Current implementation does not consider the station connections
for both the SAP's while tearing down the same but would consider
only one SAP. This leads to tearing down both the SAP's when
either of the SAP's have zero connected STA's.
This change addresses the same by considering the STA
connections on both the SAP's.

Change-Id: I53b1082c8508261e50e4e0b87be91260608c5eb7
CRs-Fixed: 2353632