aware: handle onDestroyed callback in WifiAwareStateManager context.
Aware and P2P registers for WifiManager.WIFI_STATE_ENABLED broadcast
notification. On receiving this intent:
 - if p2p client is registered, P2P initialization will start.
 - Wifi Aware will try to get capabilities in following sequence:
   getAwareInterface(); queryCapabilities(); releaseAwareInterface();

Both p2p initialization and getAwareInterface will try to get hold of
one interface. In case where aware had already got an interface and
is in process of releasing it, it will lock WifiAwareNativeManager,
and would request for HalDeviceManager lock
(removeIface->dispatchAvailableForRequestListeners->lock()).
At the same time, if P2P tries to create an interface, it need to
first remove aware interface (executeChipReconfiguration->
removeIfaceInternal->dispatchDestroyedListeners->lock()). This would
trigger a callback in WifiAwareNativeManager for cleanup(onDestroyed->
awareIsDown->lock()). With the above two sequence of operations
happening in different contexts, it may happen that P2P has already
acquired a lock in HalDeviceManager and is waiting for
WifiAwareNativeManager lock. This would result in a deadlock.

The root cause of this issue is due to these API's / operations getting
triggered in different contexts. This commit avoids the contention by
serializing one of the operations. This serialization is achieved by
having the onDestroyed event processed in WiFiAwareStateManager's context.

Change-Id: Iaece3d509e88aaf4201b23dcfff885b0c3bac8da
CRs-Fixed: 2353871