Revert "wifi: Register to ACTION_SHUTDOWN to shutdown Wi-Fi"
This reverts commit 0786c04556e9cff936137ef1fcf1917ef74bff0e.
This is causing regression with IMS module.

To fix the regression, if we try to send WifiController message
to turn off wifi, we would first need to set Settings.Global.WIFI_ON
to WIFI_DISABLED . This would mean that wifi was turned off before
reboot. Hence post reboot wifi will not turn on automatically.
This would lead to another regression.

Considering the impact of this change, wifi framework is not the
correct place to monitor for shutdown indication.

Change-Id: Ie6af63ea255a2ab58ab155145fe3d5d758b5a79c
CRs-Fixed: 2401788