wifi: Handle wifi ON state while processing CMD_DELAY_DISCONNECT.
with WIFI_DISCONNECT_DELAY_DURATION set, wifi turn off due to user
wifi disable or due to airplane mode toggle would be delayed.
This is done by parking WiFiController in QcStaDisablingState from
staEnabledState and delaying the STA DISCONNECT through
CMD_DELAY_DISCONNECT. While in this state, any requests for
WiFi ON/OFF OR APM ON/OFF are deferred to get handled once the
transition happens to the next state (based on the handling of
CMD_DELAY_DISCONNECT).

While in this phase , it does happens that the user tries to toggle
Wi-Fi / APM ON again and the current behavior fails to handle the
same. It currently handles only Wi-Fi OFF case while processing
CMD_DELAY_DISCONNECT and thus the WiFi ON scenario is not handled.

Address this by having this state also consider Wi-Fi ON scenario
while processing CMD_DELAY_DISCONNECT. Moving the current state to
the respective DisabledState's will ensure to handle the messages
that were deferred while in this state.

Tests: Set DELAY_DURATION to 10 and quick toggle airplane mode on/off.
Change-Id: Ib8deaf6fde79e283ffe4e272aadedc9c08afb21f
CRs-Fixed: 2424227