ApConfigUtil: Update upper bound for 5G frequency validation
Current 5G frequency validation logic is not considering 5845 and 5865
frequencies but these frequencies are valid in some countries. Due to
this when selected AP frequency is 5845/5865, SoftAP start is failing.
To resolve this issue update upper bound for 5G frequency validation
with current maximum possible 5G frequency i.e 5865.

Change-Id: Icb773a441ec3de2ad9c47ec062db88ae36658fcd
CRs-Fixed: 2450068