SupplicantHal/HostapdHal: Wait for death
When wpa_supplicant/hostapd is being terminated, ensure that we block
for a small timeout (50 ms) to ensure that the daemon is indeed dead. This
prevents a potential race when wifi/tethering is quickly toggled & the
previous instance of the daemon is not fully terminated before the next
toggle on.

Bug: 134729435
Test: Manually toggled wifi/tethering multiple times in quick
succession.
Test: Will send for full regression testing.

Change-Id: I0cd14e22bba21a4f7fdebd8e68642bcae9739b25
CRs-fixed: 2460780
(cherry picked from commit 8a5bc9513ceb608391721198df47c1acee0af2cd)