WiFi : Fix to "WiFi RAT lost on an Emergency Call"
The ActiveModeWarden maintains the states of respective modes through
ModeActiveState.
The transition from the respective ModeActiveState's are controlled
through the API checkForAndHandleModeChange while processing each
message obtained in the current state.
When in ClientModeActiveState and the message obtained is
ModeStateMachine.CMD_START_CLIENT_MODE,the current implementation of the
API checkForAndHandleModeChange triggers the transition to
ClientModeActiveState (same state). This further triggers the exit of
its super state (ModeActiveState) which results in Wi-Fi stop.
The end result of the current behavior is a Wi-Fi restart on an ECBM
request when the Wi-Fi is already ON.

To address this specific issue, checkForAndHandleModeChange is enhanced
to avoid the transitions to the same state.

Change-Id: Id77e2a75cec43aa60c1eabae455bdc33c14a8267
CRs-Fixed: 2501982