Append 11ax-ready capability in scan results and WifiInfo
This commit is to indicate AP HE-ready capability in scan
results and WifiInfo if both DUT and AP has below capabilities.
1. AP should be HE and 8ss capable
2. DUT should be VHT and 8ss capable

To determine AP's 8SS capability, VHT ie is checked for STS to be 7.
In this case a flag is set in Wi-Fi information and HE-Ready is added
as capability in scan results.

Change-Id: Ib235e003068e4b4be9d47fb4cf1928be5f0aa496
CRs-Fixed: 2544022