ActiveModeWarden: Read WifiApConfigStore during WifiController start.
ActiveModeWarden is initialized during wifi-service start. In init
sequence ActiveModeWarden initialized first followed by WifiApConfigStore.
This is done as WifiApConfigStore depends on ActiveModeWarden object.

Since ActiveModeWarden is trying to get WifiApConfigStore object
before it is initialized, the value is set to null.

This commit is to read the WifiApConfigStore during WifiController
start() which is invoked later in the initialization sequence.

Change-Id: I5f4b97071ce4cc3b8ed321f72d9e5764823bae8c
CRs-Fixed: 2639152