HostapdHal: configure ACS channel-list based on softap band.
For Hostapd vendor hidl, parsing of channel list is done in
HostapdHal constructor. This parsing fails if wifi-service starts
before resources/configuration are loaded.

Additionally, channel list should be configured based on softap
band.

This commit moves parsing of channel list to addVendorAccessPoint()
which is called only during softap start. Also it configures channel
list based on the band for which softap is being started.

Change-Id: Ie361589a4db7527c11b447e7c75dbdfc1624f8ae
CRs-Fixed: 2648290