Add support to roam between different SSIDs
This change enables roaming between different SSIDs on below
 - Both SSIDs AKM is WPA/WPA2-PSK
 - Both SSIDs have same default gateway MAC address

Currently framework is linking SSIDs on above conditions met and there
is an infrastructure to configure such SSIDs as whitelist SSIDs to
firmware to enable roaming between them. But framework is configuring
always empty whitelist SSIDs to firmware.

This change is to enable generating whitelist SSIDs and configuring them
to firmware based on linked SSIDs of current connected SSID. This commit
adds below changes regarding this
- Upon successful network internet validation get connected SSID default
  gateway MAC address  and persist it.
- Link other SSIDs with same default gateway MAC address and configure
  the linked SSIDs configs to supplicant and firmware.

Change-Id: If68ef24a30685e0cce13a203925a2065af4787a0
CRs-Fixed: 2746291