Wifi: Initial version of Wi-Fi Quick connect feature implementation
Quick connect feature proceeds to make a connection before the scan is
completed opportunistically.

Frameworks issues first scan request with limited channel set configured
by recent connect history based alogorithm.

Frameworks issues second scan request with all channels if first scan
request didn’t result in successful connection and enables this Quick
connect feature. Below is the mechanism to do:

When a scan is started for connection, fetch the scan results periodically
and check if any of the candidate APs received in the partial scan results
are good enough for connection. If such a candidate
AP is found, issue connection to that AP after aborting the current
ongoing scan without waiting for the scan to be completed.

In the next complete scan any better network than current connected
network found then the connection to the best network get trigger.

If any external scan(from Settings UI) request from APKs is in progress,
do not abort the scan and don't proceed for connection in order
to maintain the user experience of such APKs.

Update APKs with quick intermediate scan results periodically instead of
waiting for the full scan to be completed.

The period of fetching scan results is also configurable(default : 1 sec).

The quick connect feature is configurable.

CRs-Fixed: 2802154
Change-Id: I36dbe8596c9b57798724eec6da7ecdecf4e3224c