wifi: register VendorCallback before adding VendorAccessPoint
When starting SAP, HostapdHal adds VendorAccessPoint first(which
may just trigger an asynchronous interface setup and return), then
registers VendorCallback to monitor the status.

If this asynchronous interface setup fail in a very short time,
before the VendorCallback is registered, the failure will be missed
in HostapdHal, and result in status mismatch between framework and
hostapd (It shows SAP is enabled on UI, but in fact it's not).

To fix this issue, register VendorCallback before adding
VendorAccessPoint.

Change-Id: I0ca91b186b29b99114851097a085ec50c9e33b82
CRs-Fixed: 2867748